// daisy_scout_web/db/models/pit-2025-data.tsx
export type Pit2025Data = {
  eventId: string;
  source: string | null;
  teamNumber: number;
  scoutName: string;
  driveTrain: string;
  robotWidth: string;
  robotLength: string;
  humanPlayerLoc: string | null;
  intakeGroundAlgae: boolean | null;
  intakeReefAlgae: boolean | null;
  intakeReefCoral: boolean | null;
  intakeGroundCoral: boolean | null;
  scoreTeleopL1: boolean | null;
  scoreTeleopL2: boolean | null;
  scoreTeleopL3: boolean | null;
  scoreTeleopL4: boolean | null;
  scoreTeleopProcessor: boolean | null;
  scoreTeleopNet: boolean | null;
  scoreAutoL1: boolean | null;
  scoreAutoL2: boolean | null;
  scoreAutoL3: boolean | null;
  scoreAutoL4: boolean | null;
  scoreAutoProcessor: boolean | null;
  scoreAutoNet: boolean | null;
  prefStartCenter: boolean | null;
  prefStartAllyHP: boolean | null;
  prefStartOppHP: boolean | null;
  prefStartNone: boolean | null;
  autoIntakeHPCoral: boolean | null;
  autoIntakeGroundCoral: boolean | null;
  autoMaxGamePieces: number | null;
  canShallowClimb: boolean | null;
  canDeepClimb: boolean | null;
  noClimb: boolean | null;
  notableFeatures: string;
  createdAt: Date;
  updatedAt: Date;
};